[WFCORE-6645] Move ManagementReadXmlTestCase to Domain Mode Test Suite
